// packages/octopus-fig/src/figma-plugin-api.d.ts
interface FontName {
  readonly family: string
  readonly style: string
}

type TextCase = 'ORIGINAL' | 'UPPER' | 'LOWER' | 'TITLE'

type TextDecoration = 'NONE' | 'UNDERLINE' | 'STRIKETHROUGH'
// ...
interface Constraints {
  readonly horizontal: ConstraintType
  readonly vertical: ConstraintType
}

interface ColorStop {
  readonly position: number
  readonly color: RGBA
}

interface ImageFilters {
  readonly exposure?: number
  readonly contrast?: number
  readonly saturation?: number
  readonly temperature?: number
  readonly tint?: number
  readonly highlights?: number
  readonly shadows?: number
}

interface SolidPaint {
  readonly type: 'SOLID'
  readonly color: RGB

  readonly visible?: boolean
  readonly opacity?: number
  readonly blendMode?: BlendMode
}

interface GradientPaint {
  readonly type: 'GRADIENT_LINEAR' | 'GRADIENT_RADIAL' | 'GRADIENT_ANGULAR' | 'GRADIENT_DIAMOND'
  readonly gradientTransform: Transform
  readonly gradientStops: ReadonlyArray<ColorStop>

  readonly visible?: boolean
  readonly opacity?: number
  readonly blendMode?: BlendMode
}

interface ImagePaint {
  readonly type: 'IMAGE'
  readonly scaleMode: 'FILL' | 'FIT' | 'CROP' | 'TILE'
  readonly imageHash: string | null
  readonly imageTransform?: Transform // setting for "CROP"
  readonly scalingFactor?: number // setting for "TILE"
  readonly rotation?: number // setting for "FILL" | "FIT" | "TILE"
  readonly filters?: ImageFilters

  readonly visible?: boolean
  readonly opacity?: number
  readonly blendMode?: BlendMode
}

type Paint = SolidPaint | GradientPaint | ImagePaint
// ...
interface LetterSpacing {
  readonly value: number
  readonly unit: 'PIXELS' | 'PERCENT'
}

type LineHeight =
  | {
      readonly value: number
      readonly unit: 'PIXELS' | 'PERCENT'
    }
  | {
      readonly unit: 'AUTO'
    }

type HyperlinkTarget = {
  type: 'URL' | 'NODE'
  value: string
}

type TextListOptions = {
  type: 'ORDERED' | 'UNORDERED' | 'NONE'
}

type BlendMode =
  | 'PASS_THROUGH'
  | 'NORMAL'
  | 'DARKEN'
  | 'MULTIPLY'
  | 'LINEAR_BURN'
  | 'COLOR_BURN'
  | 'LIGHTEN'
  | 'SCREEN'
  | 'LINEAR_DODGE'
  | 'COLOR_DODGE'
  | 'OVERLAY'
  | 'SOFT_LIGHT'
  | 'HARD_LIGHT'
  | 'DIFFERENCE'
  | 'EXCLUSION'
  | 'HUE'
  | 'SATURATION'
  | 'COLOR'
  | 'LUMINOSITY'

interface Font {
  fontName: FontName
}

export interface StyledTextSegment {
  characters: string
  start: number
  end: number
  fontSize: number
  fontName: FontName
  fontWeight: number
  textDecoration: TextDecoration
  textCase: TextCase
  lineHeight: LineHeight
  letterSpacing: LetterSpacing
  fills: Paint[]
  textStyleId: string
  fillStyleId: string
  listOptions: TextListOptions
  indentation: number
  hyperlink: HyperlinkTarget | null
}
// ...
interface BaseNodeMixin extends PluginDataMixin {
  readonly id: string
  readonly parent: (BaseNode & ChildrenMixin) | null
  name: string // Note: setting this also sets `autoRename` to false on TextNodes
  readonly removed: boolean
  toString(): string
  remove(): void

  setRelaunchData(data: { [command: string]: /* description */ string }): void
  getRelaunchData(): { [command: string]: /* description */ string }
}
// ...
interface SceneNodeMixin {
  visible: boolean
  locked: boolean
  readonly stuckNodes: SceneNode[]
  readonly attachedConnectors: ConnectorNode[]
  componentPropertyReferences:
    | {
        [nodeProperty in 'visible' | 'characters' | 'mainComponent']: string
      }
    | null
}
// ...
interface ConstraintMixin {
  constraints: Constraints
}

interface LayoutMixin {
  readonly absoluteTransform: Transform
  relativeTransform: Transform
  x: number
  y: number
  rotation: number // In degrees

  readonly width: number
  readonly height: number
  readonly absoluteRenderBounds: Rect | null
  readonly absoluteBoundingBox: Rect | null
  constrainProportions: boolean

  // applicable only inside auto-layout frames
  layoutAlign: 'MIN' | 'CENTER' | 'MAX' | 'STRETCH' | 'INHERIT'
  layoutGrow: number
  layoutPositioning: 'AUTO' | 'ABSOLUTE'

  resize(width: number, height: number): void
  resizeWithoutConstraints(width: number, height: number): void
  rescale(scale: number): void
}

interface BlendMixin extends MinimalBlendMixin {
  isMask: boolean
  effects: ReadonlyArray<Effect>
  effectStyleId: string
}
// ...
interface MinimalFillsMixin {
  fills: ReadonlyArray<Paint> | 'Mixed'
  fillStyleId: string | 'Mixed'
}

interface GeometryMixin extends MinimalStrokesMixin, MinimalFillsMixin {
  strokeCap: StrokeCap | 'Mixed'
  strokeMiterLimit: number
  outlineStroke(): VectorNode | null
  fillGeometry: VectorPaths
}

interface CornerMixin {
  cornerRadius: number | 'Mixed'
  cornerSmoothing: number
}

interface RectangleCornerMixin {
  topLeftRadius: number
  topRightRadius: number
  bottomLeftRadius: number
  bottomRightRadius: number
}

interface ExportMixin {
  exportSettings: ReadonlyArray<ExportSettings>
  exportAsync(settings?: ExportSettings): Promise<Uint8Array> // Defaults to PNG format
}
// ...
interface ReactionMixin {
  reactions: ReadonlyArray<Reaction>
}

interface DocumentationLink {
  readonly uri: string
}

interface PublishableMixin {
  description: string
  documentationLinks: ReadonlyArray<DocumentationLink>
  readonly remote: boolean
  readonly key: string // The key to use with "importComponentByKeyAsync", "importComponentSetByKeyAsync", and "importStyleByKeyAsync"
  getPublishStatusAsync(): Promise<PublishStatus>
}

interface DefaultShapeMixin
  extends BaseNodeMixin,
    SceneNodeMixin,
    ReactionMixin,
    BlendMixin,
    GeometryMixin,
    LayoutMixin,
    ExportMixin {}
// ...
interface TextSublayerNode extends MinimalFillsMixin {
  readonly hasMissingFont: boolean

  paragraphIndent: number
  paragraphSpacing: number

  fontSize: number | 'Mixed'
  fontName: FontName | 'Mixed'
  readonly fontWeight: number | 'Mixed'
  textCase: TextCase | 'Mixed'
  textDecoration: TextDecoration | 'Mixed'
  letterSpacing: LetterSpacing | 'Mixed'
  lineHeight: LineHeight | 'Mixed'
  hyperlink: HyperlinkTarget | null | 'Mixed'

  characters: string
  insertCharacters(start: number, characters: string, useStyle?: 'BEFORE' | 'AFTER'): void
  deleteCharacters(start: number, end: number): void

  getRangeFontSize(start: number, end: number): number | 'Mixed'
  setRangeFontSize(start: number, end: number, value: number): void
  getRangeFontName(start: number, end: number): FontName | 'Mixed'
  setRangeFontName(start: number, end: number, value: FontName): void
  getRangeFontWeight(start: number, end: number): number | 'Mixed'
  getRangeAllFontNames(start: number, end: number): FontName[]
  getRangeTextCase(start: number, end: number): TextCase | 'Mixed'
  setRangeTextCase(start: number, end: number, value: TextCase): void
  getRangeTextDecoration(start: number, end: number): TextDecoration | 'Mixed'
  setRangeTextDecoration(start: number, end: number, value: TextDecoration): void
  getRangeLetterSpacing(start: number, end: number): LetterSpacing | 'Mixed'
  setRangeLetterSpacing(start: number, end: number, value: LetterSpacing): void
  getRangeLineHeight(start: number, end: number): LineHeight | 'Mixed'
  setRangeLineHeight(start: number, end: number, value: LineHeight): void
  getRangeHyperlink(start: number, end: number): HyperlinkTarget | null | 'Mixed'
  setRangeHyperlink(start: number, end: number, value: HyperlinkTarget | null): void
  getRangeFills(start: number, end: number): Paint[] | 'Mixed'
  setRangeFills(start: number, end: number, value: Paint[]): void
  getRangeTextStyleId(start: number, end: number): string | 'Mixed'
  setRangeTextStyleId(start: number, end: number, value: string): void
  getRangeFillStyleId(start: number, end: number): string | 'Mixed'
  setRangeFillStyleId(start: number, end: number, value: string): void
  getRangeListOptions(start: number, end: number): TextListOptions | 'Mixed'
  setRangeListOptions(start: number, end: number, value: TextListOptions): void
  getRangeIndentation(start: number, end: number): number | 'Mixed'
  setRangeIndentation(start: number, end: number, value: number): void
  getStyledTextSegments<
    StyledTextSegmentFields extends (keyof Omit<StyledTextSegment, 'characters' | 'start' | 'end'>)[]
  >(
    fields: StyledTextSegmentFields,
    start?: number,
    end?: number
  ): Array<Pick<StyledTextSegment, StyledTextSegmentFields[number] | 'characters' | 'start' | 'end'>>
}
// ...
interface TextNode extends DefaultShapeMixin, ConstraintMixin, TextSublayerNode {
  readonly type: 'TEXT'
  clone(): TextNode

  textAlignHorizontal: 'LEFT' | 'CENTER' | 'RIGHT' | 'JUSTIFIED'
  textAlignVertical: 'TOP' | 'CENTER' | 'BOTTOM'
  textAutoResize: 'NONE' | 'WIDTH_AND_HEIGHT' | 'HEIGHT' | 'TRUNCATE'
  autoRename: boolean

  textStyleId: string | 'Mixed'
  listSpacing?: number
}